Update patch set 1
Patch Set 1:

> Noticed review while checking something else... honestly I am not
 > sure we need klish any longer ;-) so maybe we could
 > deprecate/removed it.

Are you sure? I thought that was a tool for the admins to work with our system as if it was a CISCO shell?

 > Moreover I have no ideas about klish running as a deamon 0_o
 > Why do we need it?

Me neither, but I was going over the init script and just did a conversion, and only noticed that the original init script was broken after having done the work, so I thought I might as well push it out. :)

Patch-set: 1