Update patch set 1
Patch Set 1: Code-Review-1

I think that it's better to find out whether it's needed or not, and if needed, whether it's needed as a daemon.

Because maybe nowadays sysadmins run it by hand and, not being a daemon because the init script is (said to be) broken, everyone is happy.

Whereas if you fix it, there will be a running daemon that it's not needed.

So, subverting mprokop's motto when coming from FOSDEM, in this case: Better Broken than Done! :P

Patch-set: 1
Reviewer: Manuel Montecelo <1000042@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=-1