Update patch set 1
Patch Set 1:

> I think that it's better to find out whether it's needed or not,
 > and if needed, whether it's needed as a daemon.

As I've said above, it's certainly not needed, because the init script has been broken since its addition.

 > Because maybe nowadays sysadmins run it by hand and, not being a
 > daemon because the init script is (said to be) broken, everyone is
 > happy.

AFAIUI the daemon is just acts as a cache, so speedup. Still.

 > Whereas if you fix it, there will be a running daemon that it's not
 > needed.

That's why I explicitly made it not be enabled on install, which means it will not get started on install nor boot.

Patch-set: 1
Label: Code-Review=0